requirements.txt
langchain-community
fastapi
uvicorn
pydantic
python-multipart
python-dotenv
passlib==1.7.4
bcrypt==4.0.1
python-jose
cryptography
sqlparse
langchain-experimental
presidio-analyzer
presidio-anonymizer
sentence-transformers
chromadb
pypdf
langchain-ollama
langchain-huggingface
langchain-chroma
langchain-text-splitters
langchain-core
